Moved go-vcr functions to new file for easier integration by other token providers
